refactor: continued refactorign file structure
