+++ bricks/r13n/hooks/lib/hooks.dart
@@ -1,3 +1,4 @@
+export 'process_overrides.dart';
 export 'r13n_arb_document.dart';
 export 'r13n_compatibility.dart';
 export 'r13n_configuration.dart';
