+++ bricks/r13n/hooks/test/lib/r13n_configuration_test.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:mocktail/mocktail.dart';
+import 'package:r13n_hooks/hooks.dart';
 import 'package:test/test.dart';
 
 import '../../pre_gen.dart';
@@ -40,7 +41,7 @@ template-arb-file: TEMPLATE_ARB_FILE
         () async {
           await expectLater(
             R13nConfiguration.read,
-            throwsA(isA<YamlNotFoundException>()),
+            throwsA(isA<R13nYamlNotFoundException>()),
           );
 
           verifyNever(() => configFile.readAsString());
